PyQt5 API updates are not compatible with PyQt4.
We need to warn users that PyQT4 is NOT compatible with current version of PyQT5.  This should be run in a legacy PyQT4 virtual environment.  It will take considerable effort to update this to PyQT5 because many PyQT5 interfaces are different.  See the website here for the differences between PyQT4 and PyQT5:  https://doc.bccnsoft.com/docs/PyQt5/pyqt4_differences.html